some fixes

+++ intervalmap.hpp
@@ -21,7 +21,7 @@ public:
   using value_type = std::pair<Key const, Value>;
 
   using difference_type = std::ptrdiff_t;
-  using size_type = std::size_t;
+  using size_type = std::uintmax_t;
   using reference = value_type&;
   using const_reference = value_type const&;
