+++ multiset.hpp
@@ -19,7 +19,7 @@ public:
   using value_type = Key;
 
   using difference_type = std::ptrdiff_t;
-  using size_type = std::size_t;
+  using size_type = std::uintmax_t;
   using reference = value_type&;
   using const_reference = value_type const&;
